+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added `clearInvisibleValues` property to the `Survey` class.
 * Replaces deprecated `navigationButtonsVisibility` to current `showNavigationButtons` and `navigationButtonsLocation`.
 * Added `"initial"` as a value for `questionOrder` for panels to the documentation.
+* Added `kwargs` to `surveyFromJson()` to allow passing extra arguments to the `Survey` class.
 
 ## 0.5.1.1